Solution to Find the Median of an Array

+++ Find the Median of an Array@@676c02002dbc8c3e7a280b8c/code.c
@@ -9,8 +9,8 @@ int main(){
         scanf("%d",&arr[i]);
         sum+=arr[i];
     }
-    int total=sum/n;
-    printf("%d",total);
+    float total=(float)sum/n;
+    printf("%.2f",total);
     return 0;
 
 }
